Move tests to proper spot
