+++ domain/src/test/resources/webster/learners/domainExample.json
