+++ domain/src/test/resources/webster/learners/websterExample.json
