+++ domain/src/test/resources/webster/spanish/domainAnimal.json
