+++ domain/src/test/resources/webster/spanish/domainPero.json
