+++ domain/src/test/resources/webster/spanish/websterAnimal.json
