+++ domain/src/test/scala/com/foreignlanguagereader/domain/external/definition/webster/common/WebsterHeadwordInfoTest.scala
@@ -1,4 +1,4 @@
-package com.foreignlanguagereader.api.contentsource.webster.common
+package com.foreignlanguagereader.domain.external.definition.webster.common
 
 import com.foreignlanguagereader.domain.external.definition.webster.common.WebsterHeadwordInfo
 import org.scalatest.funspec.AnyFunSpec
